requirements.txt
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
easydict==1.13
Flask==3.1.0
fonttools==4.57.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.5
onnx==1.17.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.30.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
